+++ src/frontend/Orion/Configuration/AdminSettings/NodeManagement/shared/NodeMgmtPanelItem/EditModalButton.js
@@ -6,8 +6,8 @@ import { transparentize } from 'polished'
 
 import { Colors } from 'frontend/utils/pulseStyles'
 import Modal from 'frontend/components/Modal'
-import SaveButton from 'frontend/components/BusinessObjectModal/BusinessObjectModal/ButtonGroup/SaveButton'
-import NodeForm from '../NodeForm'
+
+import EditModalContent from './EditModalContent'
 
 const EditIcon = styled(FontAwesomeIcon)({
   border: 'none',
@@ -37,28 +37,9 @@ const EditButton = styled.button({
 
 const EditModalButton = ({ node }) => {
   const [showModal, setShowModal] = useState(false)
-  const [nodeFormData, setNodeFormData] = useState(node)
-
-  // const [save] = useMutation(UPDATE_TEAM_NODE, {
-  //   variables: {
-  //     input: {
-  //       nodeData,
-  //       teamId,
-  //     }
-  //   },
-  //   refetchQueries,
-  //   onCompleted: ({ updateSourceNode }) => {
-  //     console.log(updateSourceNode)
-  //     setData(updateSourceNode)
-  //     alert('Write successful')
-  //   },
-  //   onError: e => alert(`Write failure: ${e.message}`),
-  //   awaitRefetchQueries: true,
-  // })
 
   const handleModalClose = () => {
     setShowModal(false)
-    setNodeFormData(node)
   }
 
   return (
@@ -66,13 +47,8 @@ const EditModalButton = ({ node }) => {
       <EditButton onClick={() => setShowModal(true)}>
         <EditIcon size="lg" icon={faEdit} />
       </EditButton>
-      <Modal
-        show={showModal}
-        handleClose={handleModalClose}
-        title={`Editing Source Node / ${node.name}`}
-        submitButton={<SaveButton save={() => {}} inFlight={false} />}
-      >
-        <NodeForm data={nodeFormData} setData={setNodeFormData} />
+      <Modal disableHeader show={showModal}>
+        <EditModalContent node={node} handleModalClose={handleModalClose} />
       </Modal>
     </>
   )